+++ pick-front/src/data/project/MenuData.js
@@ -32,7 +32,7 @@ export const menuItems = [
         to: '/project/issues',
         label: 'ISSUE',
         title: '이슈 목록',
-        description: 'gPalsqkqh',
+        description: '이슈 목록을 조회할 수 있습니다.',
         match: ['/project/issues'],
         type: 'page'
     },
@@ -48,7 +48,7 @@ export const menuItems = [
         to: '/project/pull-requests',
         label: 'PR',
         title: 'PR 목록',
-        description: '혜민바보',
+        description: 'PR 목록을 조회할 수 있습니다.',
         match: ['/project/pull-requests'],
         type: 'page'
     },
