--- pick-front/src/project/views/IssueListView.vue
@@ -1,3 +1,0 @@
-<template>
-    <div>이슈 목록</div>
-</template>
