+++ pick-front/src/views/project/ProjectDetailView.vue
