+++ pick-front/src/views/project/ProjectInfoView.vue
